+++ api/src/jobs/index.ts
@@ -8,6 +8,7 @@ import { startSlaBreach } from './slaBreach';
 import { startSubscriptionRenewal } from './subscriptionRenewal';
 import { startCampaignWorkflowRunner } from './campaignWorkflowRunner';
 import { startWorkflowCronSchedules } from './workflowCronBootstrap';
+import { startWebhookQueueWorker } from './webhookQueueWorker';
 
 export function startBackgroundJobs() {
     startOutboxRelay();
@@ -15,5 +16,6 @@ export function startBackgroundJobs() {
     startSubscriptionRenewal();
     startCampaignWorkflowRunner();
     startWorkflowCronSchedules();
+    startWebhookQueueWorker();
     console.log('[Jobs] All background jobs registered');
 }